Move section notes to Slate Courses namespace
