--- site-root/progress/section-data.php
@@ -1,3 +1,0 @@
-<?php
-
-\Slate\Progress\SectionTermDataRequestHandler::handleRequest();
