+++ site-root/section-data.php
@@ -1,0 +1,3 @@
+<?php
+
+\Slate\Courses\SectionTermDataRequestHandler::handleRequest();
